roundNearest()   - added cast (int).

+++ Core/Geometry/IntVector.h
@@ -265,9 +265,9 @@ inline bool doesIntersect(const IntVector& low1, const IntVector &high1, const I
 // This will round the Vector v to the nearest integer
 inline IntVector roundNearest(const Vector& v)
 {
-   int x =  (v.x() < 0) ? v.x() - 0.5 : v.x() +0.5;
-   int y =  (v.y() < 0) ? v.y() - 0.5 : v.y() +0.5;
-   int z =  (v.z() < 0) ? v.z() - 0.5 : v.z() +0.5;
+   int x =  (v.x() < 0) ? (int) (v.x() - 0.5) : (int) (v.x() +0.5);
+   int y =  (v.y() < 0) ? (int) (v.y() - 0.5) : (int) (v.y() +0.5);
+   int z =  (v.z() < 0) ? (int) (v.z() - 0.5) : (int) (v.z() +0.5);
    
    IntVector ret(x,y,z);
    return ret;
